requirements.txt
Django>=2.0,<3.0
django-graphql-jwt==0.3.0
djangorestframework==3.10.3
graphene-django==2.2.0
psycopg2>=2.7,<3.0